zap2epg: Only cleanup irrelevant ShowCache files
When switching xdetails between true or false it ends-up
removing all "unecessary" ShowCache files when set to false.
This lead to a full re-downloading of the ShowCache when resetting
to true.  The following patch aims to solve this by generating
a showList using a dedicated function genShowList() so that
only irrelevant files gets cleaned-up instead of everything.

This ensure to keep found ShowCache files between updates that are
part of the TV guide.